Reorganize mbe tests
